requirements.txt
asgiref==3.7.2
attrs==24.2.0
Babel==2.10.3
barcode==1.0.4
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==3.4.8
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-3-jet==1.0.8
django-admin-argon-dashboard==1.0.16
django-admin-autocomplete-filter==0.7.1
django-admin-material-dashboard==1.0.11
django-admin-soft-dashboard==1.0.12
django-dynamic-raw-id==3.0
django-filter==23.2
django-humanize==0.1.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-jet-4==0.0.1
django-translations==1.3.0
docopt==0.6.2
docutils==0.20.1
ebaysdk==2.1.5
et-xmlfile==1.1.0
freezegun==1.2.1
geoip2==2.9.0
humanize==4.8.0
idna==2.10
isodate==0.7.2
Jinja2==3.1.2
Levenshtein==0.25.1
libsass==0.22.0
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.2
maxminddb==2.6.2
mysqlclient==2.2.4
num2words==0.5.10
odfpy==1.4.1
ofxparse==0.21
openpyxl==3.1.2
passlib==1.7.4
pefile==2023.2.7
pep8==1.7.1
Pillow==9.4.0
platformdirs==4.3.6
polib==1.1.1
psutil==5.9.4
psycopg2==2.9.5
pycparser==2.21
pydot==1.4.2
pyOpenSSL==21.0.0
pyparsing==3.2.0
PyPDF2==2.12.1
pypiwin32==223
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.8.2
python-decouple==3.8
python-Levenshtein==0.25.1
python-stdnum==1.19
pytube==15.0.0
pytz==2024.2
pyusb==1.2.1
pywebio==1.8.3
pywin32==308
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.9.3
reportlab==3.6.12
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.2.0
segno==1.6.1
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
tablib==3.5.0
tornado==6.4
typing_extensions==4.12.2
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.5
user-agents==2.2.0
vobject==0.9.6.1
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.2.1
